+++ Changelog.md
@@ -47,6 +47,7 @@
 - Rename non manifold mesh to defective mesh since some kinds of non-manifold meshes will slice correctly
 - Sliced file diff tool
 - Mesh repair tool
+- Detect and report slicing defects (caused by mesh defects)
 
 ## v0.9.2 &mdash; August 30th, 2026